requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-huggingface==0.0.1
langchain-openai==0.0.5
langgraph==0.0.20
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.4
transformers==4.36.2
openai==1.12.0
python-dotenv==1.0.0

